Standard adaptor tests
